backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
numpy>=1.26,<2.1
Pillow>=10,<12
torch>=2.2,<2.6
torchvision>=0.17,<0.21
transformers>=4.45,<4.57
faiss-cpu>=1.8,<1.13
opencv-python>=4.10,<4.13
scikit-image>=0.24,<0.26
